Extract training commonalities.
Extracted the pieces of train() that appear in all learners.
Learners now use Template Method pattern to allow each learner to
behave in calculating accuracies, when to stop, how to analyze each
row, how to initialize weights, and how to take care of algorithm
specific implementations.

In addition, this fixes the way the Cross Validation method of
training/testing determines sizes of training/validation sets. It was
calculating the size of these based off of the whole data set
(including the test portion).

 Also modified the perceptron to check accuracies against the
 validation set instead of the training set.