Tweak Manager's ML loading.
The specific ML algorithm can be loaded without using the command
line. This can be especially helpful for testing, when you need to
set all of the weights to specific values to get deterministic
results that can be tested and verified to be the same, every time.
This will allow for building the layers independently of the
command line, which will be helpful for the lab itself.