Fix expected output calculation.
The expected output that needs to be used for error calculation
_must_ be between [0,1], because that is the range of the sigmoid
function. This means that if there are four outputs in the dataset,
the expected output for each node in calculation should only be 1
for the relevant output node, and 0 for _all_ of the other nodes in
the output layer.

The algorithm appears to be learning completely correctly now. This
was the big problem with it. All of the tests pass now, and the
datasets have reasonably high accuracy when run.

This broke some of the tests in the BackProp tester, but the Node
class tester is still working perfectly, which probably means all
is okay.